Compile for Scala 3
